% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{as.data.frame.rtrace_diagnostic_set}
\alias{as.data.frame.rtrace_diagnostic_set}
\title{Convert a diagnostic set to a data frame}
\usage{
\method{as.data.frame}{rtrace_diagnostic_set}(x, ...)
}
\arguments{
\item{x}{An \code{rtrace_diagnostic_set}.}

\item{...}{Unused.}
}
\value{
A \code{data.frame} with one row per diagnostic.
}
\description{
Convert a diagnostic set to a data frame
}
