% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ast_line_count}
\alias{ast_line_count}
\title{Number of lines in a parsed file}
\usage{
ast_line_count(ast)
}
\arguments{
\item{ast}{An \code{rtrace_file_ast}.}
}
\value{
Integer scalar.
}
\description{
Number of lines in a parsed file
}
