% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{c.rtrace_diagnostic_set}
\alias{c.rtrace_diagnostic_set}
\title{Combine diagnostic sets}
\usage{
\method{c}{rtrace_diagnostic_set}(...)
}
\arguments{
\item{...}{\code{rtrace_diagnostic_set} objects.}
}
\value{
A single combined \code{rtrace_diagnostic_set}.
}
\description{
Combine diagnostic sets
}
