% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{file_hash}
\alias{file_hash}
\title{Compute a file's content hash}
\usage{
file_hash(path)
}
\arguments{
\item{path}{Character scalar file path.}
}
\value{
Character scalar MD5 hash, or \code{NA_character_} if the file
doesn't exist.
}
\description{
Compute a file's content hash
}
