% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{find_superassignments}
\alias{find_superassignments}
\title{Find all \verb{<<-} (superassignment) usages in a parsed file}
\usage{
find_superassignments(ast)
}
\arguments{
\item{ast}{An \code{rtrace_file_ast}.}
}
\value{
A \code{data.frame} with columns \code{line1}, \code{col1}.
}
\description{
Find all \verb{<<-} (superassignment) usages in a parsed file
}
