% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{get_module}
\alias{get_module}
\title{Get a registered module by id}
\usage{
get_module(id)
}
\arguments{
\item{id}{Character scalar module id.}
}
\value{
A module list, or \code{NULL} if not registered.
}
\description{
Get a registered module by id
}
