% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_discovery.R
\name{is_rtrace_plugin}
\alias{is_rtrace_plugin}
\title{Check whether a package is an RTrace plugin}
\usage{
is_rtrace_plugin(pkg)
}
\arguments{
\item{pkg}{Character scalar package name.}
}
\value{
Logical scalar.
}
\description{
Check whether a package is an RTrace plugin
}
