% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{list_modules}
\alias{list_modules}
\title{List registered Trace Platform modules}
\usage{
list_modules()
}
\value{
A named list of module registration records.
}
\description{
List registered Trace Platform modules
}
