% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin_discovery.R
\name{list_plugin_packages}
\alias{list_plugin_packages}
\title{List all installed RTrace plugin packages}
\usage{
list_plugin_packages(lib_paths = .libPaths())
}
\arguments{
\item{lib_paths}{Character vector of library paths. Defaults to
\code{\link[=.libPaths]{.libPaths()}}.}
}
\value{
Character vector of package names.
}
\description{
List all installed RTrace plugin packages
}
