% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{new_diagnostic_set}
\alias{new_diagnostic_set}
\title{Construct a set of diagnostics}
\usage{
new_diagnostic_set(diagnostics = list())
}
\arguments{
\item{diagnostics}{A list of \code{rtrace_diagnostic} objects.}
}
\value{
An object of class \code{rtrace_diagnostic_set}.
}
\description{
Construct a set of diagnostics
}
