% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules_packageqa.R
\name{packageqa-rules}
\alias{packageqa-rules}
\title{Package QA rules}
\description{
Rules for evaluating R package metadata completeness and convention
compliance. All are prefixed \verb{packageqa.*} and run through the Package
QA engine.
}
