% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walker.R
\name{path_matches_any_glob}
\alias{path_matches_any_glob}
\title{Test whether a relative path matches any of a set of glob patterns}
\usage{
path_matches_any_glob(rel_path, globs)
}
\arguments{
\item{rel_path}{Character scalar, POSIX-style relative path.}

\item{globs}{Character vector of glob patterns.}
}
\value{
Logical scalar.
}
\description{
Test whether a relative path matches any of a set of glob patterns
}
