% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platform_version}
\alias{platform_version}
\title{Return the Trace Platform version string}
\usage{
platform_version()
}
\value{
Character scalar.
}
\description{
Return the Trace Platform version string
}
