% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read an RTrace configuration file}
\usage{
read_config(path)
}
\arguments{
\item{path}{Path to a YAML configuration file.}
}
\value{
An \code{rtrace_config} object, validated via \code{\link[=validate_config]{validate_config()}}.
}
\description{
Read an RTrace configuration file
}
