% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{relative_path}
\alias{relative_path}
\title{Convert an absolute path to a project-relative path, for diagnostics}
\usage{
relative_path(context, path)
}
\arguments{
\item{context}{An \code{rtrace_context}.}

\item{path}{Character scalar absolute path.}
}
\value{
Character scalar relative path.
}
\description{
Convert an absolute path to a project-relative path, for diagnostics
}
