% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter_console.R
\name{reporter_console}
\alias{reporter_console}
\title{Render a diagnostic set as colored console output}
\usage{
reporter_console(diagnostics, use_color = cli::num_ansi_colors() > 1)
}
\arguments{
\item{diagnostics}{An \code{rtrace_diagnostic_set}.}

\item{use_color}{Logical; whether to use ANSI color. Defaults to
\code{cli::num_ansi_colors() > 1}.}
}
\value{
Invisibly, the character string written to the console.
}
\description{
Render a diagnostic set as colored console output
}
