% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_ast_cache}
\alias{write_ast_cache}
\title{Write a project's AST cache to disk}
\usage{
write_ast_cache(root, cache)
}
\arguments{
\item{root}{Character scalar project root.}

\item{cache}{A named list as returned by \code{\link[=read_ast_cache]{read_ast_cache()}}.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Write a project's AST cache to disk
}
